% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electorAr.R
\docType{package}
\name{electorAr}
\alias{electorAr}
\title{\code{electorAr} package}
\description{
Caja de Herramientas para el procesamiento de discursos presidenciales de Argentina
See the README on
\href{https://github.com/politicaargentina/electorAr/blob/master/README.md}{Github}
}
